Start implementing Order Details screen

+++ Restaurant/Models/Order.cs
@@ -10,7 +10,7 @@ namespace Restaurant.Models
     {
         public int Id { get; set; }
         public decimal SubTotal { get; set; }
-        public int Discount { get; set; }
+        public decimal Discount { get; set; }
         public decimal PreTaxTotal { get; set; }
         public decimal Tax { get; set; }
         public decimal Total { get; set; }
